Refactor location picker
